+++ tests/test_service_providers.py
@@ -1,5 +1,6 @@
 from unittest.mock import MagicMock
 from sqlalchemy.orm import Session, sessionmaker
+from limberframework.cache.cache_service_provider import CacheServiceProvider
 from limberframework.config.config_service_provider import ConfigServiceProvider
 from limberframework.database.connections import SqliteConnection
 from limberframework.database.database_service_provider import DatabaseServiceProvider
@@ -25,3 +26,10 @@ def test_authentication_service_provider():
     auth.register()
 
     assert mock_app.bind.called_once()
+
+def test_cache_service_provider():
+    mock_app = MagicMock()
+    cache = CacheServiceProvider(mock_app)
+    cache.register()
+
+    assert mock_app.bind.call_count == 2